chore: migrate DesktopNotifications module
