+++ package.json
@@ -21,6 +21,7 @@
         "tesseract.js": "^2.1.5"
     },
     "devDependencies": {
+        "@types/greasemonkey": "^4.0.2",
         "@types/jquery": "^3.5.14",
         "@types/node": "^17.0.33",
         "ajv": "^6.12.2",
